# backend/docs/catalog/openapi.yaml
openapi: 3.0.0
info:
  title: Music Catalog Service
  version: 1.0.0
  description: Сервис для управления метаданными треков и поиска.
servers:
  - url: http://api.spotik.com/catalog
tags:
  - name: Tracks
paths:
  /tracks/search:
    get:
      summary: Поиск треков
      operationId: searchTracks
      tags:
        - Tracks
      parameters:
        - name: q
          in: query
          description: Поисковой запрос (название, артист, альбом)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Список найденных треков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tracks:
    post:
      summary: Создание метаданных трека
      operationId: createTrackMetadata
      tags:
        - Tracks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackRequest'
      responses:
        '201':
          description: Трек успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tracks/{id}:
    get:
      summary: Получить трек по ID
      operationId: getTrack
      tags:
        - Tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали трека
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '404':
          description: Трек не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        album:
          type: string
          nullable: true
        duration:
          type: integer
          description: Длительность в секундах
        coverUrl:
          type: string
          format: uri
          description: URL обложки
        audioUrl:
          type: string
          format: uri
          description: URL аудио-потока
      required:
        - id
        - title
        - artist
        - duration
        - audioUrl

    CreateTrackRequest:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: integer
        mediaId:
          type: string
          format: uuid
          description: ID загруженного аудиофайла
        coverMediaId:
          type: string
          format: uuid
          description: ID загруженной обложки
      required:
        - title
        - artist
        - duration
        - mediaId

    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message